+++ src/minecraft/powercrystals/powerconverters/power/railcraft/TileEntityRailCraftConsumer.java
@@ -26,9 +26,9 @@ public class TileEntityRailCraftConsumer extends TileEntityEnergyConsumer<IFluid
 
         if (_steamTank.getFluidAmount() > 0) {
             int amount = Math.min(_steamTank.getFluidAmount(), PowerConverterCore.throttleSteamConsumer.getInt());
-            int energy = amount * PowerConverterCore.powerSystemSteam.getInternalEnergyPerInput();
+            float energy = amount * PowerConverterCore.powerSystemSteam.getInternalEnergyPerInput();
             energy = (int) storeEnergy(energy, false);
-            int toDrain = amount - (energy / PowerConverterCore.powerSystemSteam.getInternalEnergyPerInput());
+            int toDrain = (int) (amount - (energy / PowerConverterCore.powerSystemSteam.getInternalEnergyPerInput()));
             _steamTank.drain(toDrain, true);
             _mBLastTick = toDrain;
         } else {
